Refactor and commented most of my code. Also courses are now loadable in the edit schedule view
