--- src/main/java/edu/gcc/comp350/team4project/resources/templates/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>$Title$</title>
-</head>
-<body>
-    <h1>TEAM 4 Schedule App</h1>
-    <p>Yes</p>
-</body>
-</html>
